Add a missing include and match a constant type better.  From phajdan.jr. Review URL: http://codereview.chromium.org/2922
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2300 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/common/net/cookie_monster_sqlite.cc
@@ -5,6 +5,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/common/net/cookie_monster_sqlite.h"
 
+#include <list>
+
+#include "base/basictypes.h"
 #include "base/logging.h"
 #include "base/ref_counted.h"
 #include "base/string_util.h"
@@ -105,7 +108,7 @@ void SQLitePersistentCookieStore::Backend::BatchOperation(
   // Commit every 30 seconds.
   static const int kCommitIntervalMs = 30 * 1000;
   // Commit right away if we have more than 512 outstanding operations.
-  static const int kCommitAfterBatchSize = 512;
+  static const size_t kCommitAfterBatchSize = 512;
   DCHECK(MessageLoop::current() != background_loop_);
 
   // We do a full copy of the cookie here, and hopefully just here.
